% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_script_path.R
\name{get_current_script_path}
\alias{get_current_script_path}
\title{Get the full path or filename of the currently running script}
\usage{
get_current_script_path(only_filename = FALSE, throw_error_if_missing = TRUE)
}
\arguments{
\item{only_filename}{Logical; if \code{TRUE}, only the filename (not the full path) is returned. Defaults to \code{FALSE}.}

\item{throw_error_if_missing}{Logical; if \code{TRUE}, throws an error if the path cannot be determined. If \code{FALSE}, returns \code{NULL}. Defaults to \code{TRUE}.}
}
\value{
A string with either the full normalized path or just the filename of the current script, based on \code{only_filename}.
}
\description{
This function attempts to retrieve the path of the currently active script
in RStudio or the currently sourced script in non-interactive sessions.
If neither is available, it will return \code{NULL} or throw an error based on the specified \code{throw_error_if_missing} parameter.
}
\details{
Returns \code{NULL} if the script path cannot be determined and \code{throw_error_if_missing} is set to \code{FALSE}.
}
\examples{
get_current_script_path() # Returns full path if available
get_current_script_path(only_filename = TRUE) # Returns only the filename of current file
}
