% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScriptFilePath.R
\name{get_script_file_path}
\alias{get_script_file_path}
\title{Get the file path of the currently running script}
\usage{
get_script_file_path()
}
\value{
A string with the full normalized path of the current script.
}
\description{
This function attempts to retrieve the path of the currently active script
in RStudio or the currently sourced script in non-interactive sessions.
If neither is available, it will throw an error.
}
