% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_custom.r
\name{ggsave_custom}
\alias{ggsave_custom}
\title{Custom ggsave function to save ggplot figures with metadata in filename}
\usage{
ggsave_custom(
  name,
  filepath = "../06_graphics/",
  dateformat = "\%y\%m\%d",
  ...
)
}
\arguments{
\item{name}{A string representing the name of the plot (to be included in the filename).}

\item{filepath}{The path where the plot should be saved. Defaults to "../06_graphics/".}

\item{dateformat}{The date format to be used in the filename. Defaults to "\%y\%m\%d".}

\item{...}{Additional arguments passed to ggsave (e.g., width, height, etc.).}
}
\value{
Saves a file and returns the full path of the saved file.
}
\description{
This function saves a ggplot object with a custom filename that includes
the script name and the current date. It uses a flexible naming convention
to ensure better traceability of figures.
}
\details{
An error if the file path does not exist and cannot be created.
}
