% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Helper function to save ggplot or base R plots}
\usage{
save_plot(full_path, plot_obj = NULL, print_from_device = FALSE, ...)
}
\arguments{
\item{full_path}{The full file path where the plot will be saved.}

\item{plot_obj}{The plot object for ggplot2 (if applicable).}

\item{print_from_device}{A logical indicating whether to save the plot from the current device (for base R plots).}

\item{...}{Additional arguments passed to the ggsave function (for ggplot) or png (for base R).}
}
\value{
The full path of the saved file.
}
\description{
This function saves a plot (either ggplot2 or base R) to a file, based on the provided options.
}
