% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_figure_save_path.R
\name{set_figure_save_path}
\alias{set_figure_save_path}
\title{Helper function to set figure_save_path in .Rprofile}
\usage{
set_figure_save_path(path = NULL, interactive = TRUE, confirm_overwrite = TRUE)
}
\arguments{
\item{path}{Optional. A character string representing the directory to save figures. Defaults to NULL.
If NULL, the user is prompted interactively to input a path. You can use '../yourpathname' to go up one level.}

\item{interactive}{Logical. Whether to run the function interactively. Defaults to TRUE.}

\item{confirm_overwrite}{Logical. Whether to ask for confirmation before overwriting an existing .Rprofile file. Defaults to TRUE.}
}
\value{
The path that was set in the \code{.Rprofile} file.
}
\description{
This function helps users set and persist the \code{figure_save_path} option in the project's \code{.Rprofile}.
Users can specify a directory to store figues and choose whether to apply the setting interactively.
}
\examples{

library(gecko.utils)

# Set the figure_save_path to interactively
set_figure_save_path()
}
