% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{show_dataframe_in_viewer}
\alias{show_dataframe_in_viewer}
\title{Show Dataframe in Viewer}
\usage{
show_dataframe_in_viewer(df, ...)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{...}{Additional arguments to pass to \code{reactable::reactable}.}
}
\description{
Show a data frame in a viewer
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
