% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gecko.R
\name{theme_gecko}
\alias{theme_gecko}
\title{Custom theme for ggplot2 with font setup and font fallback}
\usage{
theme_gecko(
  font_family = "Noto Sans",
  title_size = 20,
  subtitle_size = 14,
  caption_size = 9,
  axis_title_size = 12,
  axis_text_size = 11
)
}
\arguments{
\item{font_family}{A string representing the font family to use. Defaults to "Noto Sans".
Fallbacks to "Arial" or "sans" if not found.}

\item{title_size}{Font size for the plot title. Defaults to 20.}

\item{subtitle_size}{Font size for the plot subtitle. Defaults to 14.}

\item{caption_size}{Font size for the plot caption. Defaults to 9.}

\item{axis_title_size}{Font size for the axis titles. Defaults to 12.}

\item{axis_text_size}{Font size for the axis text. Defaults to 11.}
}
\value{
A ggplot2 theme object.
}
\description{
This function provides a custom ggplot2 theme with options for changing fonts and text sizes.
It ensures that extrafont is loaded and the fonts are registered if necessary.
}
