% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_latex.R
\name{ggsave_latex}
\alias{ggsave_latex}
\title{Save ggplot figure for LaTeX workflow with file version management}
\usage{
ggsave_latex(
  plot_name,
  plot_obj = NULL,
  base_dir = getOption("figure_save_path", "./"),
  latex_subdir = "latex",
  archive_subdir = "archive",
  date_format = "\%y\%m\%d",
  print_from_device = FALSE,
  ...
)
}
\arguments{
\item{plot_name}{A string representing the name of the plot (to be included in the filename).}

\item{plot_obj}{The ggplot object to be saved.
If not provided, it will use the last ggplot created (\code{ggplot2::last_plot()}).}

\item{base_dir}{The base path for the plot. Defaults to current working directory. You can use set_ggsave_path() to change the default path for your total R-project.}

\item{latex_subdir}{The subdirectory for LaTeX files within the base path. Defaults to "latex".}

\item{archive_subdir}{The subdirectory where older versions are moved. Defaults to "archive".}

\item{date_format}{The date format to be used for archiving. Defaults to "\%y\%m\%d".}

\item{print_from_device}{A logical value indicating whether to print the plot from the device (base R plot) or from ggplot2. Defaults to FALSE.}

\item{...}{Additional arguments passed to ggsave (e.g., width, height, etc.).}
}
\value{
The full path of the saved figure in the LaTeX directory.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

The functionality of this function has been superseded by the \code{save_plot_with_metadata()} function.

This function saves ggplot figures to a directory used for LaTeX, ensuring that only the newest
version of each figure (without a date in the filename) is stored. Older versions are moved to an archive.
}
